api/requirements.txt
fastapi
uvicorn
pymongo
python-decouple

